I am unable to get OgreKit to compile under Leopard. So temporarily I added a precompiled framework and removed the instruction to compile OgreKit from XCode. Of course the source code is still in the SVN texshop project.
This version uses the 10.4 libraries and thus works fine on Tiger. But it also compiles on Leopard.

I upped the version number to 2.12-svn (because I privately made 2.11 and gave it to a couple of people).  Otherwise this is unchanged from 2.10-svn. But bigger changes are coming in the next few days.



git-svn-id: https://svn.code.sf.net/p/texshop/code/trunk@259 8c80bfd2-15fc-0310-8b7a-ba317743b8bb